Made prepareFractionForCyclesExport method in Analysis
